Expose dictionary utils in dictionary.h
